🔨 refactor: int -> long 변경

+++ src/main/java/com/x1/frans/supplier/query/controller/SupplierQueryController.java
@@ -39,7 +39,7 @@ public class SupplierQueryController {
 
     @Operation(summary = "공급처 상세 조회", description = "공급처 상세 조회합니다.")
     @GetMapping("/detail/{supplierId}")
-    public ResponseEntity<SupplierDetailDTO> detail(@PathVariable("supplierId") int supplierId) {
+    public ResponseEntity<SupplierDetailDTO> detail(@PathVariable("supplierId") long supplierId) {
 
         SupplierDetailDTO detail = supplierQueryService.getSupplierDetail(supplierId);
